[feat/#22] refactor: 2주차 PR 리뷰 반영
