+++ src/main/java/com/airdnb/clone/domain/filter/service/FilterService.java
@@ -8,10 +8,12 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Slf4j
 @RequiredArgsConstructor
 @Service
+@Transactional(readOnly = true)
 public class FilterService {
 
     private final BookingRepository bookingRepository;
