+++ src/main/java/com/airdnb/clone/domain/stay/controller/dto/request/edit/DescriptionEditRequest.java
@@ -1,14 +1,11 @@
 package com.airdnb.clone.domain.stay.controller.dto.request.edit;
 
 import jakarta.validation.constraints.NotBlank;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.RequiredArgsConstructor;
 
 @Getter
-@NoArgsConstructor(access = AccessLevel.PRIVATE, force = true)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@RequiredArgsConstructor
 public class DescriptionEditRequest {
 
     @NotBlank
