+++ src/main/java/com/airdnb/clone/domain/stay/controller/dto/request/edit/RoomInfoEditRequest.java
@@ -4,14 +4,11 @@ import com.airdnb.clone.domain.common.Guest;
 import com.airdnb.clone.domain.stay.entity.RoomInformation;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.RequiredArgsConstructor;
 
 @Getter
-@NoArgsConstructor(access = AccessLevel.PRIVATE, force = true)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@RequiredArgsConstructor
 public class RoomInfoEditRequest {
 
     @NotNull
